+++ tests/ws_scaffold.py
@@ -1,12 +1,5 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the Apache 2.0 License.
-import os
-import getpass
-import time
-import logging
-import multiprocessing
-import shutil
-from random import seed
 import infra.ccf
 import infra.proc
 import infra.notification
@@ -21,21 +14,16 @@ from loguru import logger as LOG
 @reqs.supports_methods("getReceipt", "verifyReceipt", "LOG_get")
 @reqs.at_least_n_nodes(2)
 def test(network, args, notifications_queue=None):
-    primary, backup = network.find_primary_and_any_backup()
+    primary, _ = network.find_primary_and_any_backup()
 
-    with primary.node_client() as mc:
-        check_commit = infra.checker.Checker(mc, notifications_queue)
-        check = infra.checker.Checker()
-
-        msg = "Hello world"
-
-        LOG.info("Write/Read on primary")
-        # TODO: input frame should have path, body and authorise field
-        # output frame should have index, term, return code and body
-        with primary.user_client(ws=True) as c:
-            for i in range(1, 51):
-                r = c.rpc("LOG_record", {"id": 42, "msg": msg * i})
-                assert r.data == b'true\n'
+    msg = "Hello world"
+    LOG.info("Write/Read on primary")
+    # TODO: input frame should have path, body and authorise field
+    # output frame should have index, term, return code and body
+    with primary.user_client(ws=True) as c:
+        for i in range(1, 51):
+            r = c.rpc("LOG_record", {"id": 42, "msg": msg * i})
+            assert r.data == b"true\n"
 
     return network
